requirements.txt
# ── Azure Identity ───────────────────────────────────────────────────
azure-identity>=1.19.0

# ── API ──────────────────────────────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.10.0
requests>=2.31.0

# ── Utilities ────────────────────────────────────────────────────────
python-dotenv>=1.0.0

# ── Testing ──────────────────────────────────────────────────────────
pytest>=8.0.0
httpx>=0.27.0
